refactor(MakeMyCard): create subfolders for   components, context and hooks for better maintainance
