+++ src/app/(pages)/fathers-and-figures/MakeMyCard/components/inputStory/SubmitStoryButton.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import SubmitDrawer from './SubmitDrawer'
+import SubmitDrawer from '../submitDrawer/SubmitDrawer'
