+++ src/app/(pages)/fathers-and-figures/MakeMyCard/components/inputStory/WriteMyCard.tsx
