+++ src/app/(pages)/fathers-and-figures/MakeMyCard/components/submitDrawer/SubmitDrawer.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React from 'react'
+import React, { use } from 'react'
 import {
     Drawer,
     DrawerClose,
@@ -11,11 +11,14 @@ import {
     DrawerTrigger,
   } from "@/components/ui/drawer"
 import { Button } from "@/components/ui/button"
-import StoryCanvasCard from './StoryCanvasCard'
+import StoryCanvasCard from '../cardBackground/StoryCanvasCard'
 import { DownloadIcon, SendIcon, ArrowLeftIcon } from "lucide-react"
+import { useAuthContext } from '@/auth/useAuthContext'
 
 
 const SubmitDrawer = () => {
+
+    const { isAuthenticated, userName} = useAuthContext();
     
   return (
     <Drawer>
@@ -34,6 +37,7 @@ const SubmitDrawer = () => {
                     <DrawerTitle></DrawerTitle>
                     <DrawerDescription></DrawerDescription>
                 </DrawerHeader>
+                <p className='text-center !p-0 !text-xs'>Hello {userName}, you're {isAuthenticated ? 'verified.' : 'not verified.'}</p>
                 <div className='flex justify-center items-center w-full text-container '>
                     <input className='w-full text-xs p-1 border border-[var(--primary-blue)] rounded-sm' type="email" placeholder='Verify your email' disabled />
                 </div>
