+++ src/auth/types.ts
@@ -1,0 +1,7 @@
+
+export interface AuthContextType {
+    isAuthenticated: boolean;
+    userName?: string;
+}
+
+
